+++ src/functions.php
@@ -1,23 +1,24 @@
 <?php
 
+use Izzle\Translation\ParameterEnclosure;
 use Izzle\Translation\Services\Translation;
 
 if (!function_exists('trans')) {
     /**
      * @param string $key
-     * @param mixed $default
+     * @param array $parameters
      * @return mixed
      * @throws Exception
      */
-    function trans($key, $default = null)
+    function trans($key, array $parameters = [])
     {
-        return Translation::translate($key, $default);
+        return Translation::translate($key, $parameters);
     }
 }
 
 if (!function_exists('trans_load')) {
-    function trans_load($json_file)
+    function trans_load($json_file, ParameterEnclosure $enclosure, $prefix = '')
     {
-        Translation::load($json_file);
+        Translation::load($json_file, $enclosure, $prefix);
     }
 }