+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 0.1.0
 -----
-- Removed default value at translation method
-- Removed bootstrap file
-- Removed default parameter
 - Added replacing parameters in translation strings
+- Added key prefix
+- Removed autoloader and debug stuff
+- Removed default parameter
 
 0.0.1
 -----